refactor: update event bus handling and imports for consistency
