--- src/handler/__init__.py
@@ -1,2 +1,0 @@
-from incoming_telegram_messages_handler import IncomingTelegramMessagesHandler
-from send_telegram_messages_handler import SendTelegramMessagesHandler
