+++ src/handler/incoming_telegram_messages_handler.py
@@ -1,18 +1,20 @@
 import logging
 
 from entity.dto import TelegramMessageDTO
-from interface import EventType, EventHandler, EventBus
+from interface.event_bus import EventBus
+from interface.event_handler import EventHandler
+from interface.event_type import EventType
 
 
 class IncomingTelegramMessagesHandler(EventHandler):
     def __init__(self):
-        self._bus = None
+        self._event_bus = None
         self._logger = logging.getLogger(self.__class__.__name__)
 
-    def init(self, bus):
-        self._bus = bus
+    def init(self, event_bus: EventBus):
+        self._event_bus = event_bus
 
-        bus.subscribe(EventType.incoming_telegram_message, self.handle_incoming_telegram_message)
+        event_bus.subscribe(EventType.incoming_telegram_message, self.handle_incoming_telegram_message)
         return True
 
     async def handle_incoming_telegram_message(self, message: TelegramMessageDTO):
@@ -24,4 +26,4 @@ class IncomingTelegramMessagesHandler(EventHandler):
             text=f"got: {message.text}",
             username=None
         )
-        self._bus.publish(EventType.send_telegram_message, replay)
+        self._event_bus.publish(EventType.send_telegram_message, replay)
