--- src/utils/__init__.py
@@ -1,1 +1,0 @@
-from event_bus import event_bus
